requirements.txt
aiodataloader==0.4.0
aniso8601==7.0.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
graphene==3.4.3
graphene-sqlalchemy==3.0.0rc1
graphql-core==3.2.5
graphql-relay==3.2.0
graphql-server==3.0.0b7
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
promise==2.3
python-dateutil==2.9.0.post0
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
